separate everythingi into individual files
