+++ lib/fields/due_date_field.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+import 'task_field.dart';
+
+
+class DueDateField extends TaskField {
+  DateTime _dueDate;
+  TimeOfDay _dueTime;
+
+  DueDateField({
+    required DateTime dueDate,
+    required TimeOfDay dueTime,
+  })  : _dueDate = dueDate,
+        _dueTime = dueTime,
+        super(
+            name: 'Due Date',
+            value: '${_formatDate(dueDate)} ${_formatTime(dueTime)}');
+
+  DateTime get dueDate => _dueDate;
+  set dueDate(DateTime value) {
+    _dueDate = value;
+    updateValue();
+  }
+
+  TimeOfDay get dueTime => _dueTime;
+  set dueTime(TimeOfDay value) {
+    _dueTime = value;
+    updateValue();
+  }
+
+  void updateValue() {
+    super.value = '${_formatDate(_dueDate)} ${_formatTime(_dueTime)}';
+  }
+
+  static String _formatDate(DateTime date) {
+    return '${date.year}-${date.month.toString().padLeft(2, '0')}-${date.day.toString().padLeft(2, '0')}';
+  }
+
+  static String _formatTime(TimeOfDay time) {
+    return '${time.hour.toString().padLeft(2, '0')}:${time.minute.toString().padLeft(2, '0')}';
+  }
+}
+
+// TODO: make it so we don't need t copy and paste these in each file
+
+String _formatDate(DateTime date) {
+  return '${date.year}-${date.month.toString().padLeft(2, '0')}-${date.day.toString().padLeft(2, '0')}';
+}
+
+String _formatTime(TimeOfDay time) {
+  return '${time.hour.toString().padLeft(2, '0')}${time.minute.toString().padLeft(2, '0')}';
+}
