+++ lib/fields/task_field.dart
@@ -1,0 +1,9 @@
+class TaskField {
+  final String name;
+  String value;
+
+  TaskField({
+    required this.name,
+    required this.value,
+  });
+}
